server/requirements.txt
Flask==2.3.2
numpy==1.24.2
Flask-Cors==4.0.0
pymongo==4.5.0
Pillow==9.5.0
matplotlib==3.8.0